requirements_dev.txt

pyelftools
pyyaml
vlsim
cocotb
pybfms

-e git+ssh://git@github.com/fvutils/pyhvl-rpc.git#egg=pyhvl-rpc

-e git+ssh://git@github.com/pybfms/pybfms.git#egg=pybfms
-e git+ssh://git@github.com/pybfms/pybfms-generic-sram.git#egg=pybfms-generic-sram
-e git+ssh://git@github.com/pybfms/pybfms_mem_core.git#egg=pybfms_mem_core
-e git+ssh://git@github.com/pybfms/pybfms-core-debug-common.git#egg=pybfms-core-debug-common
-e git+ssh://git@github.com/pybfms/pybfms_riscv_debug.git#egg=pybfms_riscv_debug

-e git+ssh://git@github.com/fvutils/mkdv.git#egg=mkdv

